requirements.txt
streamlit
pandas
numpy
plotly
fpdf2
xlsxwriter
